refactor: factor out config implementations for conditional compilation
