+++ proxy/src/config_env.rs
@@ -1,0 +1,6 @@
+use anyhow::Result;
+use std::env;
+
+pub fn read(name: &str) -> Result<String> {
+    Ok(env::var(name)?)
+}
